requirements.txt
attrs==20.3.0
certifi==2020.11.8
click==7.1.2
click-plugins==1.1.1
cligj==0.7.0
fastapi==0.61.2
Fiona==1.8.17
geopandas==0.8.1
gower==0.0.5
h11==0.11.0
joblib==0.17.0
munch==2.5.0
numpy==1.19.4
pandas==1.1.4
psycopg2-binary==2.8.6
pydantic==1.7.2
pyproj==3.0.0.post1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
scikit-learn==0.23.2
scipy==1.5.4
Shapely==1.7.1
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.20
starlette==0.13.6
threadpoolctl==2.1.0
uvicorn==0.12.2
